[ngsolve] move ngsolve wrappers to bindings.ngsolve
